feat(client): extract git operations to a separate module
* - Create a new module `git_ops.rs` and move git related operations from `client/mod.rs` to the new module
* - Implement `GitOps` trait for `Client` in `git_ops.rs`
* - Update `client/mod.rs` to import and expose `GitOps`
* - Update `lib.rs` to expose `GitOps`
* - Update `main.rs` to import `GitOps`
