+++ src/lib.rs
@@ -4,5 +4,6 @@ mod pr_title;
 mod release_notes_provider;
 
 pub use client::Client;
+pub use client::GitOps;
 pub use error::Error;
 pub use pr_title::PrTitle;
